[MODORDERS-1108] Split PieceDeleteFlowManager and create PieceDeleteFlowInventoryManager
