+++ src/main/java/org/folio/service/pieces/flows/delete/PieceDeleteFlowManager.java
@@ -1,23 +1,15 @@
 package org.folio.service.pieces.flows.delete;
 
 import static org.folio.orders.utils.ProtectedOperationType.DELETE;
-import static org.folio.service.inventory.InventoryItemManager.ITEM_STATUS;
-import static org.folio.service.inventory.InventoryItemManager.ITEM_STATUS_NAME;
-
-import java.util.Optional;
 
 import org.apache.commons.lang3.BooleanUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.folio.models.ItemStatus;
 import org.folio.models.pieces.PieceDeletionHolder;
 import org.folio.rest.RestConstants;
 import org.folio.rest.core.exceptions.ErrorCodes;
 import org.folio.rest.core.exceptions.HttpException;
 import org.folio.rest.core.models.RequestContext;
-import org.folio.rest.jaxrs.model.Piece;
 import org.folio.service.CirculationRequestsRetriever;
 import org.folio.service.ProtectionService;
 import org.folio.service.inventory.InventoryItemManager;
@@ -26,31 +18,27 @@ import org.folio.service.pieces.PieceUpdateInventoryService;
 import org.folio.service.pieces.flows.BasePieceFlowHolderBuilder;
 
 import io.vertx.core.Future;
-import io.vertx.core.json.JsonObject;
 
 public class PieceDeleteFlowManager {
 
   private static final Logger logger = LogManager.getLogger(PieceDeleteFlowManager.class);
 
+  private final PieceDeleteFlowInventoryManager pieceDeleteFlowInventoryManager;
   private final PieceStorageService pieceStorageService;
   private final ProtectionService protectionService;
-  private final InventoryItemManager inventoryItemManager;
-  private final PieceUpdateInventoryService pieceUpdateInventoryService;
   private final PieceDeleteFlowPoLineService pieceDeleteFlowPoLineService;
   private final BasePieceFlowHolderBuilder basePieceFlowHolderBuilder;
   private final CirculationRequestsRetriever circulationRequestsRetriever;
 
-  public PieceDeleteFlowManager(PieceStorageService pieceStorageService,
+  public PieceDeleteFlowManager(PieceDeleteFlowInventoryManager pieceDeleteFlowInventoryManager,
+                                PieceStorageService pieceStorageService,
                                 ProtectionService protectionService,
-                                InventoryItemManager inventoryItemManager,
-                                PieceUpdateInventoryService pieceUpdateInventoryService,
                                 PieceDeleteFlowPoLineService pieceDeleteFlowPoLineService,
                                 BasePieceFlowHolderBuilder basePieceFlowHolderBuilder,
                                 CirculationRequestsRetriever circulationRequestsRetriever) {
+    this.pieceDeleteFlowInventoryManager = pieceDeleteFlowInventoryManager;
     this.pieceStorageService = pieceStorageService;
     this.protectionService = protectionService;
-    this.inventoryItemManager = inventoryItemManager;
-    this.pieceUpdateInventoryService = pieceUpdateInventoryService;
     this.pieceDeleteFlowPoLineService = pieceDeleteFlowPoLineService;
     this.basePieceFlowHolderBuilder = basePieceFlowHolderBuilder;
     this.circulationRequestsRetriever = circulationRequestsRetriever;
@@ -64,7 +52,7 @@ public class PieceDeleteFlowManager {
       .compose(aVoid -> basePieceFlowHolderBuilder.updateHolderWithTitleInformation(holder, requestContext))
       .compose(aVoid -> protectionService.isOperationRestricted(holder.getTitle().getAcqUnitIds(), DELETE, requestContext))
       .compose(aVoid -> isDeletePieceRequestValid(holder, requestContext))
-      .compose(aVoid -> processInventory(holder, requestContext))
+      .compose(aVoid -> pieceDeleteFlowInventoryManager.processInventory(holder, requestContext))
       .compose(pair -> updatePoLine(holder, requestContext))
       .compose(aVoid -> pieceStorageService.deletePiece(holder.getPieceToDelete().getId(), true, requestContext));
   }
@@ -87,14 +75,6 @@ public class PieceDeleteFlowManager {
       .mapEmpty();
   }
 
-  private Future<Pair<String, String>> processInventory(PieceDeletionHolder holder, RequestContext requestContext) {
-    return deleteItem(holder, requestContext)
-      .compose(aVoid -> holder.isDeleteHolding()
-          ? pieceUpdateInventoryService.deleteHoldingConnectedToPiece(holder.getPieceToDelete(), requestContext)
-          : Future.succeededFuture()
-      );
-  }
-
   protected Future<Void> updatePoLine(PieceDeletionHolder holder, RequestContext requestContext) {
     var comPOL = holder.getOriginPoLine();
     return BooleanUtils.isTrue(comPOL.getIsPackage()) || BooleanUtils.isTrue(comPOL.getCheckinItems())
@@ -102,25 +82,4 @@ public class PieceDeleteFlowManager {
       : pieceDeleteFlowPoLineService.updatePoLine(holder, requestContext);
   }
 
-  private Future<Void> deleteItem(PieceDeletionHolder holder, RequestContext requestContext) {
-    var piece = holder.getPieceToDelete();
-    return getOnOrderItemForPiece(piece, requestContext)
-      .compose(item -> Optional.ofNullable(item)
-        .map(mItem -> inventoryItemManager.deleteItem(piece.getItemId(), true, requestContext))
-        .orElse(Future.succeededFuture()));
-  }
-
-  private boolean isItemWithStatus(JsonObject item, String status) {
-    return Optional.ofNullable(item)
-      .map(itemObj -> itemObj.getJsonObject(ITEM_STATUS))
-      .filter(itemStatus -> status.equalsIgnoreCase(itemStatus.getString(ITEM_STATUS_NAME)))
-      .isPresent();
-  }
-
-  private Future<JsonObject> getOnOrderItemForPiece(Piece piece, RequestContext requestContext) {
-    return StringUtils.isEmpty(piece.getItemId())
-      ? Future.succeededFuture()
-      : inventoryItemManager.getItemRecordById(piece.getItemId(), true, requestContext)
-          .map(item -> isItemWithStatus(item, ItemStatus.ON_ORDER.value()) ? item : null);
-  }
 }
